% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FKAmoli.R
\name{FKAmoli}
\alias{FKAmoli}
\title{Multi-omics integration for toxicology data}
\usage{
FKAmoli(int_method = c("sPLS-DA", "MOFA", "WGCNA", "SNF"))
}
\arguments{
\item{int_method}{sPLS-DA, MOFA2, WGCNA, or SNF}
}
\value{
output
}
\description{
Multi-omics integration for toxicology data
}
