% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_MO.R
\name{import_MO}
\alias{import_MO}
\title{Import, format and option to preprocess omics data for integration}
\usage{
import_MO(
  rnaseq_counts = NULL,
  metab_peaks = NULL,
  rrbs_mvals = NULL,
  mirna_counts = NULL,
  pirna_counts = NULL,
  meta = NULL,
  norm = TRUE,
  batch = NULL
)
}
\arguments{
\item{rnaseq_counts}{Data matrix of RNAseq counts}

\item{metab_peaks}{Data matrix of metabolite peaks}

\item{rrbs_mvals}{Data matrix of RRBS m-values}

\item{mirna_counts}{Data matrix of microRNA counts}

\item{pirna_counts}{Data matrix of piRNA counts}

\item{meta}{Metadata, must have sample ids in first column matching those in omics data and treatment col called TRT, if batch correction is required must also have unique batch column for each omics layer with batches.}

\item{norm}{Normalization, True or False}

\item{batch}{Omics layers with batches corresponding to "batch_omic" columns in metadata, column labels as follows: batch_rna, batch_metab, batch_rrbs}
}
\value{
formatted data_list for integration
}
\description{
Import, format and option to preprocess omics data for integration
}
