% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFKAmoli.R
\name{importFKAmoli}
\alias{importFKAmoli}
\title{Import, format and option to preprocess omics data for integration}
\usage{
importFKAmoli(
  rnaseq_counts = NULL,
  metab_peaks = NULL,
  rrbs_mvals = NULL,
  mirna_counts = NULL,
  pirna_counts = NULL,
  meta = NULL,
  norm = TRUE
)
}
\arguments{
\item{rnaseq_counts}{Matrix of raw RNAseq counts}

\item{metab_peaks}{Matrix of metabolite peaks}

\item{rrbs_mvals}{Matrix of RRBS m-values}

\item{mirna_counts}{Matrix of microRNA counts}

\item{pirna_counts}{Matrix of piRNA counts}

\item{meta}{Metadata, must have sample ids in first column matching those in omics data and treatment col called TRT}

\item{norm}{Normalization, True or False}
}
\value{
formatted data_list for integration
}
\description{
Import, format and option to preprocess omics data for integration
}
