% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_MO.R
\name{integrate_MO}
\alias{integrate_MO}
\title{Multi-omics integration for toxicology data}
\usage{
integrate_MO(int_method = c("sPLS-DA", "MOFA", "WGCNA", "SNF", "iPCA"))
}
\arguments{
\item{int_method}{sPLS-DA, MOFA2, WGCNA, or SNF}
}
\value{
output
}
\description{
Multi-omics integration for toxicology data
}
