% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_meta}
\alias{test_meta}
\title{Metadata for example data}
\format{
\subsection{\code{test_meta}}{

A data frame of 56 rows and 2 columns:
\describe{
\item{\code{sample}}{sample name}
\item{\code{TRT}}{treatment group M/F, Liver, treatment level}
}
}
}
\usage{
test_meta
}
\description{
Subset of fathead minnow EE2 exposure metadata collected by MIB (Janine Fetke et al.)
}
\keyword{datasets}
