% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_rnaseq_counts}
\alias{test_rnaseq_counts}
\title{RNAseq counts example data}
\format{
\subsection{\code{test_rnaseq_counts}}{

A data frame of raw RNAseq counts with 500 gene features (rows) and 56 samples (columns).
}
}
\usage{
test_rnaseq_counts
}
\description{
Subset of fathead minnow EE2 exposure RNAseq data collected by MIB (Janine Fetke et al.).
}
\keyword{datasets}
