% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_RRBS_feature_map}
\alias{test_RRBS_feature_map}
\title{RRBS feature map for example data}
\format{
\subsection{\code{test_RRBS_feature_map}}{

A data frame matching fathead minnow scaffold location to gene name
}
}
\usage{
test_RRBS_feature_map
}
\description{
Allows matching fathead minnow scaffold location to gene name
}
\keyword{datasets}
