% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_rrbs_mvals}
\alias{test_rrbs_mvals}
\title{RRBS M Values example data}
\format{
\subsection{\code{test_rrbs_mvals}}{

A data frame of RRBS M Values with 5,000 feautres (rows) and 56 samples (columns)
}
}
\usage{
test_rrbs_mvals
}
\description{
Subset of fathead minnow EE2 exposure RRBS data collected by MIB (Janine Fetke et al.)
}
\keyword{datasets}
